Collects from data from the samples
Now it tries to collect as much data from the movies. This means that
if a movie has no director, it can still have 4 actors. So extract as
much data as possible out of that data sample (take the actors), and
still use that data for ranking the actors and other stuff. However, in
the end when it’s time to write out a CSV with the cleaned data, don’t
write any movies that e.g. don’t have a director or 4 actors.